servo: adjust max output calculations
While the value is documented as ranging from -32767 to 32767, it
actually ranges from -32000 to 32000.

Additionally, the controller appears to be doing this calculation to
decide whether to increase output:
abs(output) < max_output

If we add 1 to the value returned from the controller, it consistently
comes out as being equal to the set max output.